Added support for const references
If a reference count is 0, then it will not be modified. This allows
const strings to be allocated entirely in read only memory. Because
the reference pointer is distinct from the actual struct pointers,
indirect vars such as tables and functions can have single const
instances.

Just make sure to never have more than 2,147,483,647 references to
a single object.