Changed structure of ranges to support an additional 16bit offset
Will cost an additional bitwise and operation any time a
table array is accessed, but will save memory when creating
ranges in the form `2..3`. Most importantly string slices of the
form `"hello"[2..3]` are now trivial to implement with zero
heap overhead in constant time.

Also added key and val specific "macros" in the form of inline
functions. May change the mess that is var.h to something similar
if I can figure out how to do the constant string allocation in an
inlined function. (gcc can optimize out __builtin_strlen so this
may just be possible. The inlining benifit may be significantly
platform dependent. Currently the `({ })` syntax used is gcc specific
so this is really no loss.)