Improved speed of hash functions
`str_hash` made an unescessary subtraction previously used to
designate a special string at index 0.

`num_hash` has been modified to use a magic number hack for casting
between doubles and ints. This removes all casts to integers and
allows for a decrease in instructions due to reuse of the integer.

Improvements to `str_hash` were nil, but according to gprof,
`num_hash` is now roughly twice as fast.