Moved table check for ro into tblp functions
Table pointers are passed around almost as much as actual
variables. Moving the check for const into the tlbp functions
makes it garunteed to be checked at the expense of some performace
loss due to increased checks. Since Bad Things(TM) will happen
if the tbl_t* is dereferenced with the lowest bit set it is very
important for this to be checked at every access.