Cleaned up garbage collecting semantics
The actual destruction of vars is internal to the garbage collector.
No one, in either C or V needs to access the destructors of types
directly. This lets us simplify the interface a bit and just pass
around void pointers instead of entire vars.

Note that metatables may wrap the destruction functionality, but
this can be handled in the `mtbl_destroy` function.