Made several semantic changes
Changed `null` to `nil`. Nil is more semantically accurate as the
value actually represents the complete lack of value. Because of the
difference between nil in V and null in C, which is the language V
is implemented in, the term nil was selected to emphasize this
difference and prevent mistakes from confusing the two, such as
in table assignments, which delete the values and expose values
down the tail chain.

Changed `tbl_assign` to `tbl_insert` and `tbl_set` to `tbl_assign`.
`tbl_set` and `tbl_assign` are very ambigous semantically. If you
think of this table construct as a list of tables, `tbl_insert` and
`tbl_assign` make more sense and conveying their actual functions.

Changed `off` to `pos` in a vstate struct, because thats how its used.