Modified lexing strategy for keywords
Now using lookup table for all opcodes and keywords.
Tokens are stored as bitfields in nums, this does mean tokens need
to be checked for validity, but opcodes only need to be looked up once.

Currently the table is just generated at runtime, may
compile it by hand in the future.