Changed `nil` to a literal and added `if` primary
Rather than introducing ternary operators, if
statments can be used in primary expressions.
    let a = if (b) c else d

This can be extended to other expressions
    let a = while(c) next()
    let b = for(i : 0..100) i ** 2

There are some odd definitions such as an `if` with no `else`.
    let a = if(b) c

For lack of better implementation, this is defined to assign nil
if b is false. This requires the virtual machine to handle nil values,
which can't be stored in the table of constants. A simple solution was
to add the opcode `VNIL`, which simply puts `nil` on the stack. After
this addition there is no reason to make `nil` a keyword.

This doesn't sit very well with me, but there is currently no reason
to change the nil keyword.