Cleaned up parsing/lexing code
Now the lexer stores tokens implicitely allowing the removal
of some helper functions in the parser.

Additionally length checking is performed on the generated bytecode.
This limits the length of functions significantly, however from a
few simple tests it seems the length of the bytecode roughly equals
the length of the source code. This already puts a 16 bit limitation
on the input, so putting a 16 bit limitation on the output shouldn't
be too impeding and should help catch potential bugs. Also, the 16 bit
jump limitation is no longer a concern.