Created setjmp based error handling
Any function that can create an error requires a
`veh_t *' exception handler type to be passed in.
This is currently mapped directly to a jmp_buf to allow
exceptions to occur without the overhead of checking
return values.

This does have the overhead both syntactically and runtime
of needing to pass on the argument. However, it does avoid
issues related to global state and interrupt contexts or
multithreading. Additionally, the explicit nature will
hopefully remind C users when exceptions can occur.

Note: Currently this has not been extended to the parser.
That may require additional functionality to be implemented cleanly