Renamed the programming language to Mu
As with most latin letters, there is already a programming
language named V. V was the temporary name until something else
was though of.

The name Mu was thought of for the programming language due to the
target platforms being microcontrollers, sometimes written a uc using
the greek character mu. Mu is also the japanese character for none.
Entertainingly, it is also the sound a cat makes. It is also a
convenient name in that it only uses two letters and is not confusing
to pronounce.

Additionally some naming things were cleaned up. An underscore is
used whenever Mu is used as a prefix for consistency and some var
functions have dropped the var prefix for cleanliness when they
probably won't clutter the namespace. A single M character is used
when cleaner or being merged with another prefix.