Simplified error handling by intentionally allowing memory leaks
Error handling is very annoying to implement correctly, especially
in a language without true garbage collection that uses C's own
stack for execution. However, in a correctly functioning program,
true errors should not be expected. To simplify the design of the
language and assumption made about the C runtime, Mu will simply
ignore memory leaks created when errors are propagated.

To handle errors you will be able to register an error handler in
the Mu runtime, or simply check the return of a call to the pcall
function. No memory will be cleaned up and if the program continues
this will be lost forever. This has the nice side effect of
significantly simplifying error handling and memory management.

To make this scheme for useful, I may add a voluntary garbage
collection function in the future.