Fixed an issue in error handling related to nested functions
For longjmp to work correctly, the setjmp call must not be in a
function that has returned. This is pretty unfortunate for writing
a wrapper function and requires a macro. This has been wrapped up
in an exception like macro which leaves room for other mechanics
such as c++ exceptions.