Changed all variables to fit in 32 bits (or wordsize)
After much thought (actually quite a bit), I have decided to switch
the size of variables from 64 bits to 32 bits. Using 64 bit variables
had several neat benifits, such as fitting a full double into a variable
and storing both the function pointer and closure as two pointers in a
single variable. However, 64 bit variables are large, wasted space for
nils and table pointers, and led a significant amount of complexity
just to handle these variables on the C side. For a scripting language
targeted at microcontrollers, 64 bit variables were a bad idea, if for
the space consumption alone. Switching to 32 bit variables halves the
space taken up by tables, and lets variables be stored in a single
word on most machines and eases register allocation.

Dropping 64 bit variables has three big downsides:
- The range of a number is significantly reduced. This is unavoidable.
  However, Mu has consistently used 16 bit lengths in its structures.
  Reducing this huge disparity between lengths and numbers is probably
  a good thing.

- String lengths and offset are no longer encoded in variables. The idea
  behind this was to increase data sharing between strings. However,
  this can also be achieved with string interning in a much more
  abstract and aggressive manner and string interning has the additional
  benefit of completely removing string access in common use cases

- Function closures must be stored in the function struct. This means
  a full function struct must be allocated when creating a function in
  Mu. This is a downside, but function creation still requires less
  allocations than a simple function call (which is currently the
  biggest downside to Mu).

Unrelatedly, it was noticed that reference counts don't necessarily
need to be 32 bits long. Previously 0 was chosen as a magic value
indicating that a variable was not reference counted for globally
allocated variables. This leads to a very very nice behaviour for
reference counts when they overflow: the variables simply become
constant, and reference counting is no longer performed. This may
cause a few memory leaks, but allows reference counts to be much
smaller than originally thought. 16 bit reference counts, along
with a few more changes allows both tables and functions to fit
very nicely in 4 words. This encredably convenient for the planned
arena allocation scheme. This size is now defined as MU_MINALLOC.

The most interesting thing about a 32 bit Mu, is that variables now fit
in the word size of most machines these days. There's no reason we can't
just double the size of every type for 64 bit machines. This has the
very nice benifit of improving the range for both numbers and lengths.
This does lead to a large increase in space consumption, but for 64 bit
machines, this is generally not the primary concern. To prepare for
this, 7 types have been added which form the basis for the definitions
of every other type:
[u]int_t  - Word sized integer, size must match size of pointers
[u]inth_t - Half-word sized integer
[u]intq_t - Quarter-word sized integer
data_t    - Unsigned, single character sized entry, may or may
            not be equivalent to uintq_t

It should be noted the code-base is quite a mess right not, but most
of Mu is working, albiet inefficiently due to the lack of interning,
among other things.