Moved eh pointers from the stack to thread local storage
From a theoretical point of view, passing error handlers
by parameter results in safer code, since the possibility of
an error occuring is very noticable. Passing ehs on the stack
also takes care of multithreading and portability issues.
However, errors can occur practically anywhere in Mu code, and
the syntactic overhead in C became very tedius, with wonderful
lines like this:

  tbl_insert(keyt, mcstr("nil", eh), muint(T_NIL, eh), eh);

Moving eh pointers to globals results in much cleaner code,
especially since Mu's focus on error handling has been
significantly reduced. To handle multithreading issues, a
thread-local variable is used instead of a raw global. This has the
potential to not be very portable, however due to the reliance on
sjlj, I suspect err.c won't be very portable anyways. If
necessary, err.c can also fallback to posix calls using a global
flag for initialization. Error handling doesn't need to be efficient,
so this should be fine. This is also only an issue for multithreaded
code, which is not really a big concern for Mu.

It should also be noted that there is a rather interesting issue with
the current use of sjlj. It seems that setjmp and longjmp pass 32 bit
ints, even in x86_64 mode! I'm not sure why this is, since I believe
the implementation of longjmp uses %rax for return values anyways!
The inability to pass a full pointer is very restricting and breaks
the current implementation of error handling. A work around is to
once again use a thread-local variable for passing this pointer.