Cleaned up parsing code a bit and added upper stack limit
Remove redundant encode/insert/size functions for opcode generation and
switched all arguments to be a more vm independent uint type.

Additionally, added stack counting for an upper limit on stack usage.
Operators really trip up this estimation due to the uncertainty of
if an operator will occur until an expression is completely parsed.
This would be fixed by an AST, but I am planning on changing the
language's handling of function calls which would change this anyways.