Modified base structure of language to use flexible tuples of variables
There is now a strong distinction between ',',':' and ';','='.
'a:1,b:2,c:3' now represents a tuple of the pairings of a,b,c and
1,2,3 while 'a=1;b=2;3=c' represents the assignment of a,b,c to
those values.

Function calling and table creation is now simply a modification of
these tuples. Additionally, tuples can be placed on the left side
of an assignment for flexible unpacking semantics. Tuples mostly
replace what expressions were used for and remove the asymmetry of
the calling convention.

Some examples:
a,b,c = 1,2,3
a:a, b:b, c:c = a:1, b:2, c:3
f(1,2,3)
f(a:1, b:2, c:3)
a,b,c = f(1,2,3)
a:a,b:b,c:c = f(a:1,b:2,c:3)