Converted to register based virtual machine
It should be noted that Mu does not, and will probably never get an
optimizing compiler. Originally, this was the motivation for sticking
to the simpler stack based virtual machine.

However, as the language developed, especially when tuples were added,
the number of moves needed to use the instruction set grew significantly.
The biggest contributors are the lookup/insert/assign set of
instructions for table manipulations. Since only one argument fit in an
instruction, these opcodes needed to be in static positions. Either many
wasteful moves (with ref counting) needed to be generated, or the number
of possible table instruction grew very large.

With a register based virtual machine, operand positions is a complete
non-issue. In Mu, where table operations drive the language, this is a
very big benifit. Even with the increased size of instruction, the
reduction in moves led to more compact bytecode.

Register-based instruction encodings:
[- 4 -|- 4 -|- 4 -|- 4 -]
[ op  | rd  | ra  | rb  ]
[ op  | rd  |    imm    ]

In addition to the reduced moves, a register based bytecode removes
the need to keep track of a stack pointer, which is currently simulated
on top of C's stack and does add a bit of overhead. It also removed the
need to treat a function's table scope seperated and could just store
the scope in r0 at all times in a function's execution.

Implementing the registered based bytecode was relatively
straightforward except for fitting all the operations into a good
word-size. In the end 16-bits was selected after several tradeoffs.
Only 16 registers can be available, which shouldn't be a significant
issue since most variables remain in the scope table, and large tuples
actually get stored in implicit tables when the frame size is exceeded.
Addiitonally, only 16 opcodes are available while the expected set of
needed opcodes was 17 (a very annoying number). To fit, the jump opcode
was removed. It's operation can be fabricated by a jtrue (jump if
register is true) which can use r0 (always stores the scope) for an
unconditional jump at the cost of possible performance penalty. I
suspect with all of the indirect function calls in Mu, this is not
the biggest of concerns. The full opcode listing can be found in ops.h.

Unfortunately, implementing register based bytecode in the current
parser turned out to be very difficulty. Take the following valid
statement:

a:a, b = a:a, b

Before reaching the '=', we will have no idea if we are unpacking or
packing the tuples. Previously this was handled by chunks of heap
allocated bytecode that was generated and then pieced together after
the whole statement was parsed, but with register based bytecode, we
will need to know the register indices before emitting bytecode, or
patch the bytecode which is very difficult without removing the
seperation of bytecode implementation from Mu.

The current solution is to simply use unlimited lookahead in the parser.
While this may lead to exponential runtimes for extremely ugly pieces of
Mu code, it should be sufficient for most statements. Reparsing the
statements may even be faster than allocating chunks due to reduced heap
usage and potential for the code to remain in caches.

Interestingly, the allocation of registers is actually handled just like
a stack and uses an internally tracked stack pointer to determine which
registers to use.