Implemented basic operator and call parsing
Unfortunately, setting up function calls is relatively annoying
in Mu's virtual machine. A function call takes up to MU_FRAME
arguments in registers, the function itself, and a byte indicating
the passed frame and expected frame. Due to the number of arguments,
the call opcodes can not benifit from the register nature of the
machine without exceeding the 16-bit size of instructions. The current
implementation just indicates the beginning of the call opcode's
arguments (including the actual function) in rd.

This raises a seemingly trivial question, should the function be on
the left side of the arguments or on the right. The simpliest solution
I could come up with was to put the functions on the left, and move the
first argument of operators when the operator is parsed. This does add
extra overhead to operators (especially without a move instruction),
but it keeps stack usage low and prevents non-constant passes on
expressions during parsing.

When possible, indirect references are lookuped into the correct
position, which avoids the move completely.