Added call tracking in the parser
Collected several flags into a generally state enum for tracking
what has been emitted in expressions. This allows resolution of
expressions to be deferred until the parser knows what it's doing
with them. This leads to quite a bit of nice code reuse for
unpacking/packing and lets some moves be skipped. For calls this
is expecially nice since suddenly emitting tail calls and unpacking
calls is free.