Completed structure of frame level parsing
Most of the basic stack handling has been put in place.
Nested functions are now present in the parser, and get emitted
correctly in the code structs. Files can also be parsed now,
making the parser a bit easier to test.

Additionally, the virtual machine was slightly modified. Table
manipulating opcodes have been reduced to just lookup, insert, and
assign which all operate without consuming the table, and lookdn
for consuming the table in a lookup. This effectively drops
iinsert and iassign which makes room for an unconditional jump and
register to register move. These changes were done to make the
expect_second handling for operators and binding to be less messy.
iinsert and iassign were emitted a lot less than originally
expected due to storing the scope in r0.