Merged parallel parsers into single parser with lookahead flag
More or less reverted to how the parser handled lookahead before.
Except the code has been cleaned up a bit by splitting and merging
it. Uses the p->l.lookahead flag to indicate if it is in lookahead
mode which determines if things are emitting and has some special
effects on parsing frames and things. This does cost more branches
but reduces duplication of the entire parser, which is honestly
just a bit ridiculous.