Fixed small issues with parser and introduced bracketless blocks
Now a= and =1 is no longer valid due to an explicit check for frame
sizes in assignment. This reused p_frame better than redefining
the grammar.

Originally, Mu was entirely intended to be ignorant of whitespace
and just use brackets like in C. However, this raises the issue of
where to expect brackets? C allows brackets on there own line, but
this also allows single expressions on their own line, and opens the
opportunity for surprising parses:

fn test()
{
    if (a)      # added by programmer 1 (in conditional)
        b = 2   #
        c = 2   # added by programmer 2 (not in conditional)
}

Initially, I was just going to prevent brackets from appearing on
a newline. This forces programmers to be consistent, but does limit
the expressiveness of the language a bit.

 # No newlines allowed before blocks
fn test() {
    if (a) b = 2
    c = 2
}

After toying with the grammar, it became apparent that there really
wasn't any reason to not have indention based blocks, and including
indention based blocks just made the code look nicer.

 # Indention based blocks
fn test()
    if (a)
        b = 2
        c = 2

In the end I've decided to introduction optional indention based
blocks. This is consistent with Mus handling of whitespace in
expressions and is hopefully relatively intuitive. You can still use
brackets whereever you like, which avoids python's mess with lambdas.

An initial implementation is in this commit, although it is rather
messy and scans the indention level for every space in the indention.