Seperated lexing struct from parsing struct
Lexer now only recieves a pointer to the lex struct in the parse
struct. This reduces concerns of unstored lexer state during
lookahead and seems to be a reasonable design.

Unfortunately this does inflate the size of the lex struct, which
since it's stored on the stack is actual a big concern. Later it
may be a good idea to measure stack usage and see if an alternative
method of lookahead might be a better idea.

Also if statements with basic jumps have been reintroduced.