Reduced lookahead to frame parsing
Modified parser to assume unpacking on seeing nested tables. This
prevents expressions like the following:

[a:'key'].a: val = 'key': val

But that is already a confusing usage. Much like literal keys, the
original behaviour can be forced with parenthesis:

([a:'key']).a: val = 'key': val

Additionally, in an effort to clean of the parsing code now that
lookahead usage is reduced to only frame parsing, the lookahead
has been seperated out into a set of scan functions specifically
for frames.

This removed a large amount of baggage functions that
had to perform unecessary lookahead checks and simplyfied the
effort to scan since much more of the grammar can be generalized
since full parsing is not needed. The only downside is the
duplication of some grammar rules.