Refactored lexer and removed token ascii shortcuts
Defined 'enum class' with classifications for all characters. Since
there are only 16 classes, this can be stored much more efficiently
than function pointers. Additionally, the enum constraint can allow
the compiler to more aggressively optimize switch statements/conditions.

The token enum was pruned and all ascii shortcuts were removed. They
are completely unecessary and prevent jump tables from being created
due to the enum's sparseness. To help with huge conditions the
parser had to check, the tokens were organized so that stmt/expr
classification is just a check for a range.

Additionally, most of the lexer was cleaned up. The 'mu_packed'
attribute was added specifically for hinting to c to store the
enums as tightly as possible while still tracking the possible range
of values.

Unfortunately, it looks like the mu_scan function may be difficult to
integrate.