Redefined fundamental Mu type
Rather than using a proper union to contain all possible interpretations
of the mu type, the Mu type is now defined as an opaque pointer.

This has several benefits over the union:
- The type can be used as a boolean value, which is one big benifit
  of the type's implementation. Additionally, once the num/str types
  are finished, the type will be able to be directly compared with
  equality. Kinda neat.
- Removes the spaghetti includes that were necessary to deal with
  type dependencies.
- Even on terrible compilers, the Mu type should end up in registers
  when being passed around.
- It's less likely for C users to get their grubby hands on Mu's
  internals unless they intend to.

A void * was not used to keep accidental casting from occuring.

Additionally, the Mu type was used to replace a lot of type specific
types when just passing things around. This gets rid of a whole layer
of unnecessary conversion functions and removes most of the exposed
internals of types, although it did mess up how fn/vm work together